refacto: checker namespaces
