+++ headers/builder/command-validation/checkers/Checkers.hpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   Checkers.hpp                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: csweetin <csweetin@student.42.fr>          +#+  +:+       +#+        */
+/*   By: cdomet-d <cdomet-d@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/11 15:49:12 by csweetin          #+#    #+#             */
-/*   Updated: 2025/03/20 16:38:19 by csweetin         ###   ########.fr       */
+/*   Updated: 2025/03/26 14:21:20 by cdomet-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,19 +17,12 @@
 #include "CmdSpec.hpp"
 class CmdSpec;
 
-bool hasChanPriv(CmdSpec &cmd);		// Charlotte
+bool privileges(CmdSpec &cmd);		// Charlotte
 bool isRegistered(CmdSpec &cmd);	// Charlotte
-bool joinChanRequest(CmdSpec &cmd); // Charlotte
+bool assessRequest(CmdSpec &cmd); // Charlotte
 bool onChan(CmdSpec &cmd);			// Charlotte
 bool pwMatch(CmdSpec &cmd);			// Charlotte
 bool RegStageDone(CmdSpec &cmd);	// Charlotte
-bool validChan(CmdSpec &cmd);		// Coralie
-bool validInvite(CmdSpec &cmd);		// Coralie
-bool validKick(CmdSpec &cmd);		// Coralie
-bool validMess(CmdSpec &cmd);		// Coralie
-bool validMode(CmdSpec &cmd);		// Coralie
-bool validNick(CmdSpec &cmd);		// Coralie
-bool validTarget(CmdSpec &cmd);		// Coralie
-bool validUser(CmdSpec &cmd);		// Coralie
+
 
 #endif
