+++ headers/server/Server.hpp
@@ -6,7 +6,7 @@
 /*   By: cdomet-d <cdomet-d@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/18 15:25:50 by aljulien          #+#    #+#             */
-/*   Updated: 2025/03/21 14:24:04 by cdomet-d         ###   ########.fr       */
+/*   Updated: 2025/03/26 10:12:00 by cdomet-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 #define SERVER_HPP
 
 #include "Channel.hpp"
-#include "formatMess.hpp"
+#include "n_formatMess.hpp"
 #include "typedef.hpp"
 #include <arpa/inet.h>
 #include <cstring>
